refactor(reports): one presentation layer, for every element
The first pass covered only the beam and left three flat modules at the top
level. Section 3 of the roadmap already prescribed plots/ and reports/ as
packages, so they are packages now, and every element got the same treatment.

  mento/reports/   tables.py  views.py  documents.py  walls.py  summaries.py
  mento/plots/     sections.py  walls.py  punching.py

  beam.py                2135 -> 1134   notebook views moved out
  shear_wall.py           557 ->  352   drawing, views, Word reports
  punching.py             339 ->  184   the perimeter drawing
  beam_summary.py         786 ->  588   the summary Word report
  shear_wall_summary.py   361 ->  315   idem

The notebook views moved for the same reason as the Word reports: rendering a
result to Markdown and rendering it to .docx are the same job in two media, and
neither is part of the element. That was the gap left by the previous commit.

tests/test_architecture_boundaries.py now asserts by AST that no element module
imports matplotlib, python-docx or IPython. TYPE_CHECKING imports are allowed,
since a return annotation costs nothing at runtime. Verified by adding a real
import to slab.py and watching the guard fail.

Coverage caught a behaviour change I introduced in the previous commit. To
satisfy mypy I had declared _md_properties on the beam, which silently made
`if not hasattr(self, "_md_properties")` in the results view always false - so
`beam.results` stopped printing the section line. The guard had never
short-circuited before, because `data` sets _md_data and nothing ever set
_md_properties. The call is unconditional now and the vestigial attribute is
gone.

results.py is deliberately NOT split. It holds Formatter, TablePrinter and
DocumentBuilder - the presentation primitives, which the roadmap already lists
as part of this layer. Splitting it only pays off alongside lazy imports in the
elements, and that combination buys a one-time ~600 ms of interpreter start-up
(matplotlib 514 ms, docx 104 ms) and nothing per check. mako's cost is per
check. The criterion that matters - `import mento.codes` free of both - holds
and is guarded.

Suite 959 to 965 passed, coverage still 100%, mypy clean across 45 modules.
