+++ mento/punching.py
@@ -4,13 +4,12 @@ import math
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Literal, Optional
 
-import matplotlib.pyplot as plt
-import matplotlib.patches as mpatches
 from pandas import DataFrame
 from pint import Quantity
 
 from mento.column import Column
 from mento.material import Concrete, SteelBar
+from mento.plots.punching import plot_punching_node
 from mento.units import mm, cm, inch
 
 if TYPE_CHECKING:
@@ -122,163 +121,6 @@ class Capital:
             raise ValueError("Capital thickness must be positive")
 
 
-def _plot_punching_node(node: PunchingNode) -> None:
-    """Basic plan-view plot: column, capital, openings, free edges. No critical perimeter yet."""
-    d = node.slab.d_avg.to("cm").magnitude
-    col = node.column
-
-    b_col = col.b.to("cm").magnitude
-    h_col = col.h.to("cm").magnitude if col.shape == "rectangular" else b_col
-
-    extent = max(5 * d, 3 * max(b_col, h_col), 30)
-
-    # Slab boundary (clipped by free edges where present)
-    x_slab_max = col.edge_distance_x.to("cm").magnitude if col.edge_distance_x is not None else extent
-    y_slab_max = col.edge_distance_y.to("cm").magnitude if col.edge_distance_y is not None else extent
-    x_slab_min = -extent
-    y_slab_min = -extent
-
-    fig, ax = plt.subplots(figsize=(7, 7))
-    ax.set_facecolor("white")
-
-    # Slab background — exact slab area only, RGB(227, 227, 227)
-    slab_bg = mpatches.Rectangle(
-        (x_slab_min, y_slab_min),
-        x_slab_max - x_slab_min,
-        y_slab_max - y_slab_min,
-        linewidth=0,
-        facecolor="#e3e3e3",
-        zorder=1,
-    )
-    ax.add_patch(slab_bg)
-
-    # Free edge lines — dark blue, drawn only along the slab boundary
-    _EDGE_COLOR = "#003399"
-    _EDGE_LW = 2.5
-    if col.position in ("edge", "corner"):
-        ax.plot(
-            [x_slab_max, x_slab_max],
-            [y_slab_min, y_slab_max],
-            color=_EDGE_COLOR,
-            linewidth=_EDGE_LW,
-            solid_capstyle="butt",
-            label="Free edge",
-            zorder=6,
-        )
-    if col.position == "corner":
-        ax.plot(
-            [x_slab_min, x_slab_max],
-            [y_slab_max, y_slab_max],
-            color=_EDGE_COLOR,
-            linewidth=_EDGE_LW,
-            solid_capstyle="butt",
-            zorder=6,
-        )
-
-    # Capital — clipped to the actual slab boundary
-    if node.capital is not None:
-        cap_b = node.capital.b.to("cm").magnitude
-        cap_h = node.capital.h.to("cm").magnitude
-        cx0 = max(-cap_b / 2, x_slab_min)
-        cy0 = max(-cap_h / 2, y_slab_min)
-        cx1 = min(cap_b / 2, x_slab_max)
-        cy1 = min(cap_h / 2, y_slab_max)
-        if cx1 > cx0 and cy1 > cy0:
-            cap_patch = mpatches.Rectangle(
-                (cx0, cy0),
-                cx1 - cx0,
-                cy1 - cy0,
-                linewidth=1.5,
-                edgecolor="#555555",
-                facecolor="#c0c0c0",
-                linestyle="--",
-                zorder=3,
-                label="Capital",
-            )
-            ax.add_patch(cap_patch)
-
-    # Column — medium gray (lighter than old #404040, darker than slab)
-    if col.shape == "rectangular":
-        col_patch = mpatches.Rectangle(
-            (-b_col / 2, -h_col / 2),
-            b_col,
-            h_col,
-            linewidth=2,
-            edgecolor="black",
-            facecolor="#808080",
-            zorder=4,
-        )
-    else:
-        col_patch = mpatches.Circle(
-            (0, 0),
-            b_col / 2,
-            linewidth=2,
-            edgecolor="black",
-            facecolor="#808080",
-            zorder=4,
-        )
-    ax.add_patch(col_patch)
-
-    # Openings — white fill + border + diagonal X (no hatch)
-    _OP_COLOR = "#009431"
-    for i, opening in enumerate(node.openings):
-        ox = opening.x.to("cm").magnitude
-        oy = opening.y.to("cm").magnitude
-        label = "Opening" if i == 0 else None
-
-        if opening.shape == "rectangular":
-            ob = opening.b.to("cm").magnitude
-            oh = opening.h.to("cm").magnitude
-            ax.add_patch(
-                mpatches.Rectangle(
-                    (ox - ob / 2, oy - oh / 2),
-                    ob,
-                    oh,
-                    linewidth=1.5,
-                    edgecolor=_OP_COLOR,
-                    facecolor="white",
-                    zorder=5,
-                    label=label,
-                )
-            )
-            # X from corner to corner
-            ax.plot([ox - ob / 2, ox + ob / 2], [oy - oh / 2, oy + oh / 2], color=_OP_COLOR, linewidth=1.2, zorder=6)
-            ax.plot([ox - ob / 2, ox + ob / 2], [oy + oh / 2, oy - oh / 2], color=_OP_COLOR, linewidth=1.2, zorder=6)
-        else:
-            od = opening.diameter.to("cm").magnitude
-            ax.add_patch(
-                mpatches.Circle(
-                    (ox, oy),
-                    od / 2,
-                    linewidth=1.5,
-                    edgecolor=_OP_COLOR,
-                    facecolor="white",
-                    zorder=5,
-                    label=label,
-                )
-            )
-            # X spanning the bounding box of the circle
-            r = od / 2
-            ax.plot([ox - r, ox + r], [oy - r, oy + r], color=_OP_COLOR, linewidth=1.2, zorder=6)
-            ax.plot([ox - r, ox + r], [oy + r, oy - r], color=_OP_COLOR, linewidth=1.2, zorder=6)
-
-    ax.set_xlim(-extent * 1.1, extent * 1.1)
-    ax.set_ylim(-extent * 1.1, extent * 1.1)
-    ax.set_aspect("equal")
-    ax.set_xlabel("x [cm]")
-    ax.set_ylabel("y [cm]")
-    ax.set_title(f"Punching Node — {col.position.capitalize()} {col.shape.capitalize()} Column")
-    ax.grid(True, alpha=0.3, zorder=0)
-
-    handles, labels = ax.get_legend_handles_labels()
-    if handles:
-        ax.legend(loc="upper right", fontsize=9)
-
-    plt.tight_layout()
-    plt.show()
-    plt.close("all")
-
-
 class PunchingNode:
     """
     Associates a PunchingSlab with a Column and one or more Forces objects.
@@ -323,8 +165,11 @@ class PunchingNode:
         raise NotImplementedError("Punching design not yet implemented — coming in Phase 4.")
 
     def plot(self) -> None:
-        """Display a plan-view of the punching node geometry."""
-        _plot_punching_node(self)
+        """Display a plan-view of the punching node geometry.
+
+        The drawing itself lives in :mod:`mento.plots.punching`.
+        """
+        plot_punching_node(self)
 
     def __repr__(self) -> str:
         n_forces = len(self.forces)
